requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pytest==7.4.3
httpx==0.25.2
pyjwt==2.8.0

